MINOR Moved ApiMessageAndVersion and AbstractApiMessageAndVersionSerde to clients module.
