+++ checkstyle/import-control.xml
@@ -271,7 +271,7 @@
 
     <subpackage name="log">
       <allow pkg="org.apache.kafka.server.log" />
-      <allow pkg="org.apache.kafka.metadata" />
+      <allow pkg="org.apache.kafka.server" />
     </subpackage>
   </subpackage>
 
@@ -408,6 +408,7 @@
     <allow pkg="org.apache.kafka.common.requests" />
     <allow pkg="org.apache.kafka.common.protocol" />
     <allow pkg="org.apache.kafka.metalog" />
+    <allow pkg="org.apache.kafka.server" />
     <allow pkg="org.apache.kafka.test"/>
     <allow pkg="com.fasterxml.jackson" />
     <allow pkg="net.jqwik"/>
