+++ clients/src/main/java/org/apache/kafka/server/AbstractApiMessageAndVersionSerde.java
@@ -14,12 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.server.log.remote.metadata.storage.serialization;
+package org.apache.kafka.server;
 
 import org.apache.kafka.common.protocol.ApiMessage;
 import org.apache.kafka.common.protocol.ByteBufferAccessor;
 import org.apache.kafka.common.protocol.ObjectSerializationCache;
-import org.apache.kafka.metadata.ApiMessageAndVersion;
 
 import java.nio.ByteBuffer;
