+++ metadata/src/main/java/org/apache/kafka/controller/ControllerResult.java
@@ -17,7 +17,7 @@
 
 package org.apache.kafka.controller;
 
-import org.apache.kafka.metadata.ApiMessageAndVersion;
+import org.apache.kafka.server.ApiMessageAndVersion;
 
 import java.util.Collections;
 import java.util.List;
