+++ metadata/src/main/java/org/apache/kafka/controller/ControllerResultAndOffset.java
@@ -17,7 +17,7 @@
 
 package org.apache.kafka.controller;
 
-import org.apache.kafka.metadata.ApiMessageAndVersion;
+import org.apache.kafka.server.ApiMessageAndVersion;
 
 import java.util.Objects;
 import java.util.stream.Collectors;
