+++ metadata/src/main/java/org/apache/kafka/controller/SnapshotGenerator.java
@@ -24,7 +24,7 @@ import java.util.OptionalLong;
 
 import org.apache.kafka.common.utils.ExponentialBackoff;
 import org.apache.kafka.common.utils.LogContext;
-import org.apache.kafka.metadata.ApiMessageAndVersion;
+import org.apache.kafka.server.ApiMessageAndVersion;
 import org.slf4j.Logger;
