+++ metadata/src/test/java/org/apache/kafka/controller/MockSnapshotWriter.java
@@ -18,7 +18,7 @@
 package org.apache.kafka.controller;
 
 import org.apache.kafka.common.protocol.ApiMessage;
-import org.apache.kafka.metadata.ApiMessageAndVersion;
+import org.apache.kafka.server.ApiMessageAndVersion;
 
 import java.io.IOException;
 import java.util.ArrayList;
