+++ raft/src/main/java/org/apache/kafka/raft/metadata/MetaLogRaftShim.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.raft.metadata;
 
 import org.apache.kafka.common.protocol.ApiMessage;
-import org.apache.kafka.metadata.ApiMessageAndVersion;
+import org.apache.kafka.server.ApiMessageAndVersion;
 import org.apache.kafka.metalog.MetaLogLeader;
 import org.apache.kafka.metalog.MetaLogListener;
 import org.apache.kafka.metalog.MetaLogManager;
